Tried to fix the camera, looks better but far from perfect. Also, implemented clipping in a very bad way (probably worth more ditched than fixed)TO DO: Fix the pers proj ; Allow rotation around the camera y axis

+++ libftmath/orthomat4new.c
@@ -6,7 +6,7 @@
 /*   By: lbelda <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/21 21:15:22 by lbelda            #+#    #+#             */
-/*   Updated: 2017/11/21 23:38:13 by lbelda           ###   ########.fr       */
+/*   Updated: 2017/11/23 01:28:40 by lbelda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ t_mat4	orthomat4new(t_frustrum fr)
 {
 	t_mat4	ortho;
 
-	ortho.x = vec4new(2.0 / (double)(fr.r - fr.l), 0.0, 0.0, 0.0);
+	ortho.x = vec4new(2.0 / (fr.r - fr.l), 0.0, 0.0, 0.0);
 	ortho.y = vec4new(0.0, 2.0 / (fr.t - fr.b), 0.0, 0.0);
 	ortho.z = vec4new(0.0, 0.0, -2.0 / (fr.f - fr.n), 0.0);
 	ortho.w = vec4new(-(fr.r + fr.l) / (fr.r - fr.l),
